Refactor dashboard cards
