+++ src/components/ProblemCards.test.tsx
@@ -1,0 +1,30 @@
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom';
+import ProblemCards from './ProblemCards';
+import { Difficulty } from '@/types/types';
+import type { ProblemLite } from '@/types/recommendation';
+
+describe('<ProblemCards>', () => {
+  const baseProblem: ProblemLite = {
+    slug: 'prob-1',
+    title: 'Problem One',
+    difficulty: Difficulty.Easy,
+    popularity: 0.5,
+    isFundamental: true,
+    tags: ['Array'],
+  };
+
+  it('shows tags and fundamental badge', () => {
+    render(<ProblemCards problems={[baseProblem]} bucket="fundamentals" />);
+    expect(screen.getByText('Problem One')).toBeInTheDocument();
+    expect(screen.getByText('Array')).toBeInTheDocument();
+    expect(screen.getByText('Fundamental')).toBeInTheDocument();
+  });
+
+  it('hides tags and displays last solved for refresh bucket', () => {
+    const prob = { ...baseProblem, lastSolved: Date.now() / 1000 };
+    render(<ProblemCards problems={[prob]} bucket="refresh" showTags={false} />);
+    expect(screen.queryByText('Array')).not.toBeInTheDocument();
+    expect(screen.getByText(/Last solved/)).toBeInTheDocument();
+  });
+});
